requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiofiles==23.2.1 